+++ packages/bridge/src/bridge/createBridge.ts
@@ -1,9 +1,7 @@
-import { Bridge, ErrorHandlers } from ".";
-import { BridgeConfig, IEventTypes, IRequestTypes } from "./Bridge.type";
+import { Bridge, BridgeConfig, IEventTypes, IRequestTypes } from "@/bridge";
 
 export function createBridge<T extends IRequestTypes, E extends IEventTypes>(
-  errorHandlers: ErrorHandlers,
   config: BridgeConfig
 ): Bridge<T, E> {
-  return new Bridge<T, E>(errorHandlers, config);
+  return new Bridge<T, E>(config);
 }