+++ packages/bridge/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @webviewkit/bridge
 
+## 3.0.0
+
+### Major Changes
+
+- Implement version-aware handler selection in Bridge
+- Implement version-aware handler selection in Bridge
+
 ## 2.0.0
 
 ### Major Changes